+++ anjlab-tapestry-config/src/main/java/com/anjlab/tapestry5/config/ConfigHelperModule.java
@@ -18,6 +18,7 @@ package com.anjlab.tapestry5.config;
 import com.anjlab.tapestry5.config.services.ConfigHelper;
 import com.anjlab.tapestry5.config.services.ConfigHelperInitializer;
 import com.anjlab.tapestry5.config.services.ConfigHelperValidator;
+import com.anjlab.tapestry5.config.services.PropertyTypeValidator;
 import com.anjlab.tapestry5.config.services.UnreferencedPropertiesValidator;
 import org.apache.tapestry5.ioc.OrderedConfiguration;
 import org.apache.tapestry5.ioc.ServiceBinder;
@@ -42,5 +43,6 @@ public class ConfigHelperModule
             OrderedConfiguration<ConfigHelperValidator> configuration)
     {
         configuration.addInstance("UnreferencedProperties", UnreferencedPropertiesValidator.class);
+        configuration.addInstance("PropertyType", PropertyTypeValidator.class);
     }
 }