+++ anjlab-tapestry-config/src/test/java/com/anjlab/tapestry5/config/services/ConfigHelperTest.java
@@ -26,39 +26,39 @@ public class ConfigHelperTest
     public void testFromFile() throws IOException
     {
         ConfigHelper helper = ConfigHelper.fromFile("src/test/resources/base-config.properties");
-        Assert.assertEquals("Base Property 1", helper.get("prop1"));
-        Assert.assertEquals("Base Property 2", helper.get("prop2"));
+        Assert.assertEquals("Base Property 1", helper.getRaw("prop1"));
+        Assert.assertEquals("Base Property 2", helper.getRaw("prop2"));
     }
 
     @Test
     public void testFileConfigExtendFrom() throws IOException
     {
         ConfigHelper helper = ConfigHelper.fromFile("src/test/resources/extending-config.properties");
-        Assert.assertEquals("Overridden Property 1", helper.get("prop1"));
-        Assert.assertEquals("Base Property 2", helper.get("prop2"));
+        Assert.assertEquals("Overridden Property 1", helper.getRaw("prop1"));
+        Assert.assertEquals("Base Property 2", helper.getRaw("prop2"));
     }
 
     @Test
     public void testFromClasspath() throws IOException
     {
         ConfigHelper helper = ConfigHelper.fromClasspathResource("base-config.properties");
-        Assert.assertEquals("Base Property 1", helper.get("prop1"));
-        Assert.assertEquals("Base Property 2", helper.get("prop2"));
+        Assert.assertEquals("Base Property 1", helper.getRaw("prop1"));
+        Assert.assertEquals("Base Property 2", helper.getRaw("prop2"));
     }
 
     @Test
     public void testClasspathConfigExtendFrom() throws IOException
     {
         ConfigHelper helper = ConfigHelper.fromClasspathResource("extending-config.properties");
-        Assert.assertEquals("Overridden Property 1", helper.get("prop1"));
-        Assert.assertEquals("Base Property 2", helper.get("prop2"));
+        Assert.assertEquals("Overridden Property 1", helper.getRaw("prop1"));
+        Assert.assertEquals("Base Property 2", helper.getRaw("prop2"));
     }
 
     @Test
     public void testMultiExtension() throws IOException
     {
         ConfigHelper helper = ConfigHelper.fromClasspathResource("multi-extending-config.properties");
-        Assert.assertEquals("Base Property 1", helper.get("prop1"));
-        Assert.assertEquals("Base Property 2", helper.get("prop2"));
+        Assert.assertEquals("Base Property 1", helper.getRaw("prop1"));
+        Assert.assertEquals("Base Property 2", helper.getRaw("prop2"));
     }
 }