+++ anjlab-tapestry-config/CHANGELOG.md
@@ -5,5 +5,7 @@
 
   - Upgraded to Tapestry 5.4.2
   - `ConfigHelper` package renamed to `com.anjlab.tapestry5.config.services.ConfigHelper`
+  - `ConfigHelper#names()` renamed to `ConfigHelper#getPropertyNames()`
+  - `ConfigHelper#get()` renamed to `ConfigHelper#getRaw()`
   - `anjlab-tapestry-config` is now a drop-in module that will register itself
-  - Added support for config validators, one built-in validator "UnreferencedProperties" was added
+  - Added support for config validators, built-in validators are: "UnreferencedProperties", "PropertyType"